--renamed index.js to init.js and added a placeholder for db helper functions
